@@ -1,9 +1,7 @@
 package com.codetroopers.materialAndroidBootstrap;
 
-import android.annotation.TargetApi;
 import android.app.Activity;
 import android.app.Instrumentation;
-import android.os.Build;
 import android.support.test.InstrumentationRegistry;
 import android.support.test.rule.ActivityTestRule;
 
@@ -23,13 +21,7 @@ public abstract class RecreateActivityTest<T extends Activity> {
         launchActivity();
 
         final Instrumentation instrumentation = InstrumentationRegistry.getInstrumentation();
-        instrumentation.runOnMainSync(new Runnable() {
-            @TargetApi(Build.VERSION_CODES.HONEYCOMB)
-            @Override
-            public void run() {
-                rule.getActivity().recreate();
-            }
-        });
+        instrumentation.runOnMainSync(() -> rule.getActivity().recreate());
         instrumentation.waitForIdleSync();
     }
 
