@@ -34,12 +34,7 @@ public class DrawerAdapter extends RecyclerView.Adapter<DrawerAdapter.ViewHolder
     @Override
     public void onBindViewHolder(ViewHolder viewHolder, final int position) {
         viewHolder.name.setText("menu " + position);
-        viewHolder.name.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                mListener.onClick(v, position);
-            }
-        });
+        viewHolder.name.setOnClickListener(v -> mListener.onClick(v, position));
         viewHolder.itemView.setActivated(position == mSelectedItem);
     }
 
